# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Webdata Service API
  description: |
    Servicio de scraping de productos usando Firecrawl.
    Gestiona fuentes de datos, jobs de scraping, productos scrapeados y estadisticas del pipeline.
    Multi-tenant con aislamiento por X-Tenant-ID.
  version: 1.0.0
  contact:
    name: Mercado Cercano
    url: https://mercadocercano.com
servers:
  - url: http://localhost:8150
    description: Desarrollo local (directo)
  - url: http://localhost:8001/webdata
    description: Desarrollo local (via Kong Gateway)
security:
  - BearerAuth: []
    TenantHeader: []
...
paths:
...
  /api/v1/stats:
    get:
      operationId: getStats
      summary: Estadisticas generales del pipeline
      tags: [Stats]
      parameters:
        - $ref: "#/components/parameters/XTenantID"
      responses:
        "200":
          description: Estadisticas generales
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/StatsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stats/sources:
    get:
      operationId: getSourceStats
      summary: Estadisticas por fuente
      tags: [Stats]
      parameters:
        - $ref: "#/components/parameters/XTenantID"
      responses:
        "200":
          description: Estadisticas de cada fuente
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SourceStatsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stats/pipeline:
    get:
      operationId: getPipelineStatus
      summary: Estado del pipeline por business type
      tags: [Stats]
      parameters:
        - $ref: "#/components/parameters/XTenantID"
        - name: business_type
          in: query
          required: true
          schema:
            type: string
          description: Business type para consultar el pipeline
      responses:
        "200":
          description: Estado del pipeline
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/PipelineStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por iam-service

    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: UUID del tenant
  parameters:
    XTenantID:
      name: X-Tenant-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: UUID del tenant
...
  responses:
    BadRequest:
      description: Solicitud invalida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Mensaje de error
...
    StatsResponse:
      type: object
      properties:
        total_sources:
          type: integer
        active_sources:
          type: integer
        total_jobs:
          type: integer
        pending_jobs:
          type: integer
        running_jobs:
          type: integer
        completed_jobs:
          type: integer
        failed_jobs:
          type: integer
        total_products:
          type: integer
    SourceStatsResponse:
      type: object
      properties:
        source_id:
          type: string
        name:
          type: string
        health_score:
          type: number
          format: double
        total_runs:
          type: integer
        successful_runs:
          type: integer
        failed_runs:
          type: integer
        is_active:
          type: boolean
    PipelineStatusResponse:
      type: object
      properties:
        business_type:
          type: string
        sources:
          type: object
          properties:
            total:
              type: integer
            active:
              type: integer
            healthy:
              type: integer
            failing:
              type: integer
        products:
          type: object
          properties:
            scraped:
              type: integer
            with_image:
              type: integer
            without_image:
              type: integer
            synced_to_pim:
              type: integer
            pending_sync:
              type: integer
        last_scrape:
          type: string
          format: date-time
          nullable: true
        next_scrape:
          type: string
          format: date-time
          nullable: true